This set of changes is for MELPA compliance
This includes:
1. Consistent name-spacing
2. Clean byte-compiling

Some of these changes will introduce non-backward compatible changes. One of
them is if we call the library ov-highlighter, then that is also the namespace
of functions and variables. So older files with ov-highlight-data will not be
readable here without some hackery.

In a few places I commented out some lines that appear not to be used.
Eventually, these should be deleted.